Feature: Provide custom styling for Strapi themes.
