+++ admin/src/components/CKEditorInput/GlobalStyling.js
@@ -1,133 +1,21 @@
 import {createGlobalStyle} from "styled-components";
+import { style as common } from "./styles/common";
+import { style as light } from "./styles/light";
+import { style as dark } from "./styles/dark";
 
-export const GlobalStyling = createGlobalStyle`
-  .ck-word-count {
-    display: flex;
-    gap: 0.5rem;
-    justify-content: flex-end;
-    padding-block: 0.2rem;
-    font-size: 0.875rem;
+export const getGlobalStyling = ( theme ) => {
+  let themeStyle = null;
+  switch ( theme ) {
+    case 'dark':
+      themeStyle = dark;
+      break;
+    case 'light':
+      themeStyle = light;
+      break;
   }
 
-  .ck-editor__main {
-    --ck-font-face: "Source Sans Pro", system-ui, Roboto, "Helvetica Neue", "Helvetica", Arial, sans-serif;
-    color: var(--ck-color-editor-base-text);
-    font-family: var(--ck-font-face);
-
-    * {
-      font: revert;
-      margin: revert;
-    }
-
-    h1 {
-      font-size: 2em;
-      margin-bottom: 0.6em;
-      font-weight: 400;
-    }
-
-    h2 {
-      font-size: 1.68em;
-      margin-bottom: 0.4em;
-      font-weight: 400;
-    }
-
-    h3 {
-      font-size: 1.36em;
-      margin-bottom: 0.2em;
-      font-weight: 400;
-    }
-
-    h4 {
-      font-size: 1.15em;
-      margin-bottom: 0.1em;
-      font-weight: 400;
-    }
-
-    p {
-      font-size: 1em;
-      line-height: 1.6em;
-      margin-bottom: 0.8em;
-    }
-
-    figcaption {
-      background-color: var(--ck-color-image-caption-background);
-      caption-side: bottom;
-      color: var(--ck-color-image-caption-text);
-      display: table-caption;
-      font-size: .75em;
-      outline-offset: -1px;
-      padding: 0.6em;
-      word-break: break-word;
-    }
-
-    blockquote {
-    }
-
-    .table {
-      margin: 0;
-    }
-
-    ul.todo-list {
-      list-style: none;
-      margin: revert;
-      color: revert;
-      font-family: revert;
-      margin-left: 2rem;
-    }
-
-    ul,
-    ol {
-      list-style: initial;
-      margin-left: 2rem;
-    }
-
-    ol {
-      list-style: decimal;
-    }
-
-    sub {
-      vertical-align: sub
-    }
-
-    sup {
-      vertical-align: super
-    }
-
-    // higher specificity needed
-    .ck.ck-content.ck-editor__editable {
-      line-height: initial;
-      min-height: 12.5rem;
-      border-bottom-left-radius: 0.25rem;
-      border-bottom-right-radius: 0.25rem;
-
-      // so it's more Strapi alike
-      &.ck-focused:not(.ck-editor__nested-editable) {
-        border: 1px solid rgb(73, 69, 255);
-        box-shadow: rgb(73 69 255) 0 0 0 0.125rem;
-        border-top-left-radius: 0.25rem;
-        border-top-right-radius: 0.25rem;
-        transition-property: border-color, box-shadow;
-        transition-duration: 0.2s;
-      }
-    }
-  }
-
-/*
-  // styles for dark theme
-  :root {
-    --ck-color-toolbar-background: rgb(24, 24, 38);
-    --ck-color-text: rgb(165, 165, 186);
-    --ck-color-editor-base-text: rgb(255, 255, 255);
-    --ck-color-toolbar-border: rgb(74, 74, 106);
-    --ck-color-base-border: rgb(74, 74, 106);
-    --ck-color-base-background:  rgb(33, 33, 52);
-    --ck-color-button-default-hover-background: rgb(33, 33, 52);
-    --ck-color-button-on-background: rgb(33, 33, 52);
-    --ck-color-button-on-hover-background: rgb(33, 33, 52);
-    --ck-color-button-default-active-background: rgb(33, 33, 52);
-  }
-  .ck-word-count {
-    color: rgb(165, 165, 186)
-  }
-*/
-`;
+  return createGlobalStyle`
+    ${ common }
+    ${ themeStyle }
+  `;
+};
