+++ admin/src/components/CKEditorInput/styles/dark.js
@@ -1,0 +1,19 @@
+import { css } from "styled-components";
+
+export const style = css`
+  :root {
+    --ck-color-toolbar-background: rgb(24, 24, 38);
+    --ck-color-text: rgb(165, 165, 186);
+    --ck-color-editor-base-text: rgb(255, 255, 255);
+    --ck-color-toolbar-border: rgb(74, 74, 106);
+    --ck-color-base-border: rgb(74, 74, 106);
+    --ck-color-base-background:  rgb(33, 33, 52);
+    --ck-color-button-default-hover-background: rgb(33, 33, 52);
+    --ck-color-button-on-background: rgb(33, 33, 52);
+    --ck-color-button-on-hover-background: rgb(33, 33, 52);
+    --ck-color-button-default-active-background: rgb(33, 33, 52);
+  }
+  .ck-word-count {
+    color: rgb(165, 165, 186)
+  }
+`;
