+++ admin/src/components/CKEditorInput/styles/light.js
@@ -1,0 +1,4 @@
+import { css } from "styled-components";
+
+export const style = css`
+`;
